_ Correção bugs Sonar

+++ controle-projetos/src/main/java/perrut/matheus/controleprojetos/service/impl/MemberServiceImpl.java
@@ -1,10 +1,13 @@
 package perrut.matheus.controleprojetos.service.impl;
 
+import java.util.NoSuchElementException;
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import perrut.matheus.controleprojetos.domain.Member;
 import perrut.matheus.controleprojetos.domain.Person;
+import perrut.matheus.controleprojetos.domain.Project;
 import perrut.matheus.controleprojetos.exception.InvalidEmployeeException;
 import perrut.matheus.controleprojetos.repository.MemberRepository;
 import perrut.matheus.controleprojetos.repository.PersonRepository;
@@ -26,17 +29,23 @@ public class MemberServiceImpl implements MemberService {
   @Override
   @Transactional
   public Member saveMember(Member member) {
-    projectRepository.findById(member.getProjectId()).orElseThrow();
-    Person person = personRepository.findById(member.getPersonId()).orElseThrow();
+    Optional<Project> projectOptional = projectRepository.findById(member.getProjectId());
 
-    if (!person.getEmployee()) {
-      throw new InvalidEmployeeException(person);
+    if (projectOptional.isPresent()) {
+      Person person = personRepository.findById(member.getPersonId()).orElseThrow();
+
+      if (!person.getEmployee()) {
+        throw new InvalidEmployeeException(person);
+      }
+
+      memberRepository.findById(member.getProjectId()).ifPresent(m -> {
+        memberRepository.delete(m);
+      });
+
+      return memberRepository.save(member);
+    } else {
+      throw new NoSuchElementException(
+          String.format("Project with id: %s not found", member.getProjectId()));
     }
-
-    memberRepository.findById(member.getProjectId()).ifPresent(m -> {
-      memberRepository.delete(m);
-    });
-
-    return memberRepository.save(member);
   }
 }
